Refactoring test files.
